Issue #39 feat : Notification Service bean change

+++ commons-lib/src/main/java/org/ekstep/genieservices/commons/bean/Notification.java
@@ -44,7 +44,7 @@ public class Notification implements Serializable {
         this.notificationJson = notificationJson;
     }
 
-    public double getMsgId() {
+    public int getMsgId() {
         return msgid;
     }
